LlamaIndex_UseCases/UseCase3/multi-agent-diabetes-ui/requirements.txt
llama-index==0.11.21
llama-index-llms-openai==0.2.10
tabulate==0.9.0
python-dotenv==1.0.1
openai>=1.40.0
tiktoken
streamlit==1.38.0
